Moved system of ecs into its own folder.
