+++ src/core/ecs/system/collision_entity_system.h
@@ -2,9 +2,9 @@
 #define COLLISION_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../../physics/collision/collision_context.h"
-#include "../../component/components/collider_component.h"
-#include "../../../math/space_handler.h"
+#include "../../physics/collision/collision_context.h"
+#include "../component/components/collider_component.h"
+#include "../../math/space_handler.h"
 
 class CollisionEntitySystem : public EntitySystem {
   private:
