+++ src/core/ecs/system/directional_light_rendering_entity_system.h
@@ -2,9 +2,9 @@
 #define DIRECTIONAL_LIGHT_RENDERING_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../component/components/light3D_component.h"
-#include "../../component/components/material_component.h"
-#include "../../../math/space_handler.h"
+#include "../component/components/light3D_component.h"
+#include "../component/components/material_component.h"
+#include "../../math/space_handler.h"
 
 class DirectionalLightRenderingEntitySystem : public EntitySystem {
   private:
