+++ src/core/ecs/system/point_light_rendering_entity_system.h
@@ -2,10 +2,10 @@
 #define POINT_LIGHT_RENDERING_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../component/components/transform3D_component.h"
-#include "../../component/components/light3D_component.h"
-#include "../../component/components/material_component.h"
-#include "../../../math/space_handler.h"
+#include "../component/components/transform3D_component.h"
+#include "../component/components/light3D_component.h"
+#include "../component/components/material_component.h"
+#include "../../math/space_handler.h"
 
 class PointLightRenderingEntitySystem : public EntitySystem {
   private:
