+++ src/core/ecs/system/script_entity_system.h
@@ -2,9 +2,9 @@
 #define SCRIPT_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../../scripting/script_context.h"
-#include "../../component/components/transform2D_component.h"
-#include "../../component/components/text_label_component.h"
+#include "../../scripting/script_context.h"
+#include "../component/components/transform2D_component.h"
+#include "../component/components/text_label_component.h"
 
 class ScriptEntitySystem : public EntitySystem {
   private:
