+++ src/core/ecs/system/sprite_rendering_entity_system.h
@@ -2,9 +2,9 @@
 #define SPRITE_RENDERING_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../component/components/transform2D_component.h"
-#include "../../component/components/sprite_component.h"
-#include "../../../math/space_handler.h"
+#include "../component/components/transform2D_component.h"
+#include "../component/components/sprite_component.h"
+#include "../../math/space_handler.h"
 
 class SpriteRenderingEntitySystem : public EntitySystem {
   private:
