+++ src/core/ecs/system/texture_cube_rendering_entity_system.h
@@ -2,9 +2,9 @@
 #define TEXTURE_CUBE_RENDERING_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../component/components/transform3D_component.h"
-#include "../../component/components/material_component.h"
-#include "../../../math/space_handler.h"
+#include "../component/components/transform3D_component.h"
+#include "../component/components/material_component.h"
+#include "../../math/space_handler.h"
 
 class TextureCubeRenderingEntitySystem : public EntitySystem {
   private:
