+++ src/core/ecs/system/timer_entity_system.h
@@ -2,8 +2,8 @@
 #define TIMER_ENTITY_SYSTEM_H
 
 #include "entity_system.h"
-#include "../../../timer/timer_manager.h"
-#include "../../../signal_manager.h"
+#include "../../timer/timer_manager.h"
+#include "../../signal_manager.h"
 
 class TimerEntitySystem : public EntitySystem {
   private:
