+++ src/core/scripting/python/python_modules.cpp
@@ -7,12 +7,12 @@
 #include "../../audio/audio_helper.h"
 #include "../../utils/helper.h"
 #include "../../signal_manager.h"
-#include "../../ecs/entity/system/collision_entity_system.h"
+#include "../../ecs/system/collision_entity_system.h"
 #include "../../ecs/component/components/transform3D_component.h"
 #include "../../ecs/component/components/texture_cube_component.h"
 #include "../../ecs/component/components/light3D_component.h"
-#include "../../ecs/entity/system/timer_entity_system.h"
-#include "../../ecs/entity/system/text_rendering_entity_system.h"
+#include "../../ecs/system/timer_entity_system.h"
+#include "../../ecs/system/text_rendering_entity_system.h"
 
 // ENGINE
 PyObject* PythonModules::engine_exit(PyObject *self, PyObject *args, PyObject *kwargs) {
